requirements.txt
fastapi==0.111.0
pydantic==2.7.4
alembic==1.13.2
sqlalchemy==2.0.31
uvicorn==0.30.1
asyncpg==0.29.0
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
apscheduler==3.10.4
pytz==2024.1
loguru==0.7.2
celery==5.3.4
redis~=5.0.7
gunicorn==22.0.0
starlette~=0.37.2